projects/cvs_risk/requirements.txt
-e ../../
git+https://github.com/SwanseaUniversityMedical/pyconceptlibraryclient.git@v1.0.0
pandas>=2, <3
PyYAML>=6, <7
python-dotenv